// src/user/Enums/EMessage.enum.ts
export enum EMessageFailed {
    EMAIL_IS_TAKEN = 'Email is already taken',
    USER_NOT_REGISTERED = 'User have been not registered',
    EMPTY_USER_DATA = 'User Data is empty',
    USER_HAVE_THIS_ROLE = 'User already have this role',
    USER_HAVE_THIS_USERNAME = 'User already have this username',
    USERNAME_IS_NOT_AVAIBLE = 'This username is not avaible'
}